% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_reject.R
\name{accept_reject}
\alias{accept_reject}
\title{Determines acceptance or rejection of a proposal}
\usage{
accept_reject(
  summary,
  proposal_summary,
  t,
  p_depends_delta = FALSE,
  c = 1,
  c_all = 1,
  verbose = T
)
}
\arguments{
\item{summary}{List of current objective values.}

\item{proposal_summary}{List of proposed objective values.}

\item{t}{Current temperature.}

\item{p_depends_delta}{If TRUE, acceptance depends on magnitude of worsening.}

\item{c}{Penalty multipliers. Single value of vector of same length as \code{summary}.}

\item{c_all}{Multiplier when all objectives worsen.}
}
\value{
Logical scalar indicating whether proposal is accepted.
}
\description{
Determines acceptance or rejection of a proposal
}
