% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_similar_scale}
\alias{check_similar_scale}
\title{Check trait data scaling}
\usage{
check_similar_scale(x, tolerance = 2)
}
\arguments{
\item{x}{List of trait values}

\item{tolerance}{How much variation in trait values are allowed.
Compared to the ratio of max(difference in trait values) / min(difference in trait values)}
}
\value{
An error or nothing.
}
\description{
Check trait data scaling
}
