% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_pareto.R
\name{explore_pareto}
\alias{explore_pareto}
\title{Explore Pareto Front}
\usage{
explore_pareto(
  multiopt_archive_output,
  nd_samples = 100,
  max_steps = 5000,
  verbose = T,
  ...
)
}
\arguments{
\item{multiopt_archive_output}{Unmodified archive output from \code{multiopt_sa}. out$archive}

\item{nd_samples}{How many observations of the Pareto Front to explore.
Will only impact settings for multiopt_sa internally, results may be larger than the provided \code{nd_samples}
Should probably be set higher than what you initially set for \code{multiopt_sa} for maximum Pareto front exploration.}

\item{max_steps}{Maximum number of simulated annealing iterations.
Since we are simply exploring the Pareto Front, not aiming for an optimum solution,
this value does not need to be as large as when used for the full simulated annealing run.}

\item{...}{The same arguments used to initially run \code{multiopt_sa}. If you supply \code{initial_weights} it will be overwritten.}
}
\value{
Returns a list matching the formatting of \code{multiopt_archive_output}, but with hopefully more observations of the front.
}
\description{
Expands upon the archive values from an initial simulated annealing run
to expand the known coverage of the Pareto Front.
}
\details{
Can use \code{unscale_archive} to back transform data to original trait scales.
}
\references{
Bragg, J. G., van der Merwe, M., Yap, J.-Y. S., Borevitz, J., & Rossetto, M. (2022). Plant collections for conservation and restoration: Can they be adapted and adaptable? Molecular Ecology Resources, 22(6), 2171–2182. https://doi.org/10.1111/1755-0998.13605
}
