% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weights.R
\name{modify_weights}
\alias{modify_weights}
\title{Modify weight vector}
\usage{
modify_weights(w, w_min = NULL, w_max = NULL, n_t)
}
\arguments{
\item{w}{Integer vector of current weights.}

\item{w_min}{Optional integer vector of minimum allowable weights.
May be:
\itemize{
\item NULL (defaults to 0).
\item a scalar applied to all individuals.
\item a vector of length n_g (number of individuals with trait data) indicating minimum weight for each individual.
}}

\item{w_max}{Optional integer vector of maximum allowable weights.
May be:
\itemize{
\item NULL (unconstrained)
\item a scalar applied to all individuals.
\item a vector of length n_g (number of individuals with trait data)
}}

\item{n_t}{Integer. Total target population size.}
}
\value{
Modified weight vector of same length as w.
}
\description{
Randomly selects a weight and moves it to a new individual, if possible, while preserving limitations.
}
