% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleopt_context.R
\name{singleopt_context}
\alias{singleopt_context}
\title{Run single-objective simulated annealing in context of additional traits}
\usage{
singleopt_context(
  trait_list,
  measure_list,
  measure_args_list,
  n_t = NULL,
  weights_min = NULL,
  weights_max = 1,
  initial_weights = NULL,
  max_steps = 10000,
  max_t = 1,
  p_depends_delta = F,
  acceptance_multipliers = 1,
  verbose = T,
  n_runs = 1,
  parallel = F
)
}
\arguments{
\item{trait_list}{List of trait matrices. Each matrix should contain
individuals in rows and trait values in columns.}

\item{measure_list}{List of objective functions corresponding to each
element of trait_list. Possible functions include:
\code{\link[=nei_diversity]{nei_diversity()}}, \code{\link[=shannon_diversity]{shannon_diversity()}}, \code{\link[=allele_enrichment]{allele_enrichment()}},
\code{\link[=weighted_mean_of_vector]{weighted_mean_of_vector()}}, \code{\link[=sum_of_squared_difference]{sum_of_squared_difference()}},
\code{\link[=weighted_mean_of_absolute_difference]{weighted_mean_of_absolute_difference()}},
\code{\link[=weighted_mean_of_pairwise_matrix]{weighted_mean_of_pairwise_matrix()}}, \code{\link[=wasserstein_measure]{wasserstein_measure()}}, or
\code{\link[=trait_coverage]{trait_coverage()}}. See function descriptions for specific details and
required arguments.}

\item{measure_args_list}{List containing additional arguments supplied
to each measure function.}

\item{n_t}{Integer giving the total number or total weight of individuals
to select.}

\item{weights_min}{Optional minimum allowable value(s) for weights.}

\item{weights_max}{Optional maximum allowable value(s) for weights.
Either a single numeric value or a vector with length equal to the
number of individuals. Automatically set to 1, individuals can only
be selected once.}

\item{initial_weights}{Optional vector of starting weights for the
optimization.}

\item{max_steps}{Maximum number of simulated annealing iterations.}

\item{max_t}{Initial annealing temperature.}

\item{p_depends_delta}{Logical; if \code{TRUE}, acceptance probabilities
depend on the magnitude of objective change.}

\item{acceptance_multipliers}{Numeric scalar or vector of the same length as \code{trait_list} controlling
acceptance probability for each objective.}

\item{verbose}{If TRUE, prints status updates in the console.}

\item{n_runs}{Integer. How many randomizations to run.}

\item{parallel}{Logical; whether to run in parallel. If TRUE, user will need to set up the session using \code{future::plan()}}
}
\value{
A list of the same length of \code{trait_list} with an element for each
trait found in \code{trait_list}. Within each list element, a table with measure
results for each trait is returned, with the column with the same name as the
list element being the single objective used in the simulated annealing run(s).
}
\description{
This function aims to address "If I prioritize a single trait for simulated annealing,
what are the results of the other traits?." This is complimentary analysis
to \code{multiopt_sa} and \code{rand_multiopt}, allowing users to explore what trade offs
happen when incorporating multiple objectives. No chains or archives are returned.
}
\details{
Compatible with parallel computing when n_reps > 1. See \code{rand_multiopt()} for more details.

Output value can be back transformed using \code{unscale_singleopt}.
}
\examples{

set.seed(12345)
n = 100
x = rnorm(n = n, mean = 120, sd = 2)
y = x * 3 + rnorm(n = n, mean = 0, sd = 20)
dat = data.frame(x = x, y = y)

trait_list = list(
x = as.matrix(dat$x),
y = as.matrix(dat$y)
)

measure_list = list(
 x = weighted_mean_of_vector,
 y = weighted_mean_of_vector
)

args_list = list(
 x = NULL,
 y = list(direction = -1) # to minimize the mean of this trait
)

trait_list_scaled <- scale_traits(trait_list)

set.seed(123)
singleopt_context(
 trait_list = trait_list_scaled,
 measure_list = measure_list,
 measure_args_list = args_list,
 n_t = 10,
 verbose = F
)


}
