% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_scheduler.R
\name{temp_scheduler}
\alias{temp_scheduler}
\title{Determine annealing temperature}
\usage{
temp_scheduler(s, max_steps, max_t, min_t = 0, nda = FALSE)
}
\arguments{
\item{s}{Integer. Current time step.}

\item{max_steps}{Numeric scalar. Maximum allowed time steps.}

\item{max_t}{Numeric scalar. Maximum allowed temperature.}

\item{min_t}{Numeric scalar. Minimum allowed temperature. Only relevant when nda = TRUE.}

\item{nda}{Logical. If TRUE, indicates to hold temp at minimum temperature to allow exploration of Pareto front.}
}
\value{
A numeric scalar of current annealing temperature.
}
\description{
Determine annealing temperature
}
