% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_traits.R
\name{unscale_multiopt}
\alias{unscale_multiopt}
\title{Back transform output to original scale}
\usage{
unscale_multiopt(trait_list, multiopt_output, skip_traits = NULL)
}
\arguments{
\item{trait_list}{Original list of trait data as used in \code{multiopt_sa}.}

\item{multiopt_output}{Unmodified output from \code{multiopt_sa}.}

\item{skip_traits}{Character vector of trait names that should not be
back transformed.}
}
\value{
Object of same dimensions and formatting as \code{multiopt_output}.
}
\description{
This function uses the original transformed trait data provided in \code{multiopt_sa} and the corresponding output
from \code{multiopt_sa} to back transform simulation output to the original trait scale.
This function assumes trait data was transformed using \code{scale_traits}.
}
\details{
Note that this function takes the absolute value of transformed trait variables.
}
