% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_traits.R
\name{unscale_singleopt}
\alias{unscale_singleopt}
\title{Back transform output to original scale}
\usage{
unscale_singleopt(trait_list, singleopt_output, skip_traits = NULL)
}
\arguments{
\item{trait_list}{Original list of trait data as used in \code{singleopt_context}.}

\item{singleopt_output}{Archive output from \code{singleopt_context}.}

\item{skip_traits}{Optional characted vector specifing if any traits should
be excluded from back transformation.}
}
\value{
List with archive values and weights matching formatting of \code{singleopt_output}.
}
\description{
This function uses the original transformed trait data provided to \code{singleopt_context} and the corresponding output
from \code{singleopt_context} to back transform archive values to the original trait scale.
This function assumes trait data was transformed using \code{scale_traits}.
}
\details{
Note that this function takes the absolute value of transformed trait variables.
}
