% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weights.R
\name{.check_weights}
\alias{.check_weights}
\title{Internal helper for validating weight constraints.}
\usage{
.check_weights(n_g, n_t = NULL, w_min, w_max)
}
\arguments{
\item{n_g}{Integer. Number of genotypes/individuals.}

\item{n_t}{Integer. Total target population size.}

\item{w_min}{Optional integer vector of minimum allowable weights.
May be:
\itemize{
\item NULL (defaults to 0).
\item a scalar applied to all individuals.
\item a vector of length n_g (number of individuals with trait data) indicating minimum weight for each individual.
}}

\item{w_max}{Optional integer vector of maximum allowable weights.
May be:
 \itemize{
   \item NULL (unconstrained)
   \item a scalar applied to all individuals.
   \item a vector of length n_g (number of individuals with trait data)
 }}
}
\value{
List of length 2 containing a vector of length n_g for both w_min and w_max
}
\description{
Internal helper for validating weight constraints.
}
