% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_traits.R
\name{min_max_unscale}
\alias{min_max_unscale}
\title{Reverse min-max scaling using a reference dataset}
\usage{
min_max_unscale(x_scaled, x_unscaled)
}
\arguments{
\item{x_scaled}{Numeric vector, matrix, or array containing values
scaled to the range [0, 1].}

\item{x_unscaled}{Numeric vector, matrix, or array providing the
reference scale used for back-transformation.}
}
\value{
An object with the same structure as `x_scaled`, transformed
  back to the scale of `x_unscaled`.
}
\description{
Transforms values from a min-max scaled scale back to their
original scale using the minimum and maximum of a reference
unscaled dataset.
}
\details{
The transformation is:

\deqn{
x = x_{scaled} \times (\max(x_{ref}) - \min(x_{ref})) + \min(x_{ref})
}


The minimum and maximum used for back-transformation are calculated
from `x_unscaled` with missing values ignored.

This function assumes that `x_scaled` was originally generated using
min-max scaling based on the same reference distribution.
}
