% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_figures.R
\name{plot_pareto}
\alias{plot_pareto}
\title{Visualize pairwise Pareto trade-offs}
\usage{
plot_pareto(archive_list, multi_list = NULL, single_list = NULL)
}
\arguments{
\item{archive_list}{A list containing Pareto archive output,
expected to include an element `archive_summary` with trait values.
Either  $archive from `multiopt_sa` or `rand_multiopt` or output from `explore_pareto`}

\item{multi_list}{Optional list containing multi-objective optimization
results from `rand_multiopt` or `multiopt_sa`. Must include either `measure_summaries` or
`final_selection$measure_summary`.}

\item{single_list}{Optional list of single-objective optimization
results from `singleopt_context`.}
}
\value{
A patchwork object consisting of multiple ggplot2
  pairwise scatterplots with shared legends.
}
\description{
Generates a grid of pairwise scatterplots showing trade-offs among
traits in a Pareto archive. Optionally overlays results from
multi-objective and single-objective optimization runs for comparison.
}
\details{
Required packages: ggplot2, dplyr, patchwork.
}
