% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weights.R
\name{propose_initial_weights}
\alias{propose_initial_weights}
\title{Generate initial weights}
\usage{
propose_initial_weights(n_g, n_t, w_max = NULL, w_min = NULL, verbose = FALSE)
}
\arguments{
\item{n_g}{Integer. Number of genotypes/individuals.}

\item{n_t}{Integer. Total target population size.}

\item{w_max}{Optional integer vector of maximum allowable weights.
May be:
 \itemize{
   \item NULL (unconstrained)
   \item a scalar applied to all individuals.
   \item a vector of length n_g (number of individuals with trait data)
 }}

\item{w_min}{Optional integer vector of minimum allowable weights.
May be:
\itemize{
\item NULL (defaults to 0).
\item a scalar applied to all individuals.
\item a vector of length n_g (number of individuals with trait data) indicating minimum weight for each individual.
}}

\item{verbose}{Logical. Print progress messages.}
}
\value{
Integer vector of weights
}
\description{
Generates initial weights for use before initiating simulated annealing.
}
