% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_multiopt.R
\name{rand_multiopt}
\alias{rand_multiopt}
\title{Multiple runs of simulated annealing}
\usage{
rand_multiopt(n_runs = 20, parallel = F, multiopt_args, verbose = T)
}
\arguments{
\item{n_runs}{Integer. How many randomizations to run.}

\item{parallel}{Logical; whether to run in parallel. If TRUE, user will need to set up the session using `future::plan()`}

\item{multiopt_args}{List of arguments to supply to `multiopt_sa`}

\item{verbose}{Whether status updates should be printed to the console.}
}
\value{
a list containing:
\describe{
  \item{measure_summaries}{Table containing the final objective values.}
  \item{individs_selected}{Matrix of weights for each individual (columns) for each randomization (row).}
  \item{archive}{If `nda = TRUE`, a list containing archived
  non-dominated solutions (`archive_summary`) and associated weights (`archive_weights`); otherwise `NULL`.
  Archive is the combined archives over all randomizations.}
}
}
\description{
This is a wrapper function around `multiopt_sa()` for multiple replicates
of the simulated annealing process. Results are summarized and condensed.
Process can be run in parallel using the `future` package.
}
\examples{

library(future)

future::plan(multisession, workers = 4) # multisession indicates this is run on this computer


set.seed(12345)
x = rnorm(n = n, mean = 120, sd = 2)
y = x * 3 + rnorm(n = n, mean = 0, sd = 20)
dat = data.frame(x = x, y = y)

trait_list = list(
x = as.matrix(dat$x),
y = as.matrix(dat$y)
)

measure_list = list(
 x = weighted_mean_of_vector,
 y = weighted_mean_of_vector
)

args_list = list(
 x = NULL,
 y = list(direction = -1)
)

trait_list_scaled <- scale_traits(trait_list)

sa_args = list(
trait_list = trait_list_scaled,
measure_list = measure_list,
measure_args_list = args_list,
n_t = 50,
weights_max = 1,
max_steps = 50000,
nda = T,
nd_samples = 500
)

test = rand_multiopt(n_runs = 5, multiopt_args = sa_args, parallel = T)
str(test)

}
