% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_measure.R
\name{sum_of_squared_difference}
\alias{sum_of_squared_difference}
\title{#' Compute weighted sum of squared deviations from a target value}
\usage{
sum_of_squared_difference(v, w, disp = 0, direction = 1)
}
\arguments{
\item{v}{single-column matrix of numeric values. Will be coerced to a vector.}

\item{w}{Numeric vector of individual weights with same length as `v`.}

\item{disp}{Numeric scalar target value to compare the weighted mean against. Defaults to 0.}

\item{direction}{numeric scalar. Multiplier applied to the final metric value
to control orientation for simulated annealing. Use 1 for default direction
(maximize trait value), -1 to invert the sign (minimize trait value).}
}
\value{
A single numeric value representing the weighted sum of squared
        deviations from `disp`.
}
\description{
Calculates the weighted sum of squared differences between a numeric vector
(`v`) and a displacement value (`disp`). Each squared deviation is weighted
by a corresponding value in `w`.This could be used, e.g., to minimize
the mean difference between temperature of origin for each sample,
and temperate of a site.
}
\details{
The function computes:
\deqn{ \sum_i w_i (v_i - disp)^2 }

Direction is applied after metric computation and does not alter the
underlying metric definition.
}
\examples{
v <- as.matrix(c(1, 2, 3))
w <- c(0.2, 0.3, 0.5)
sum_of_squared_difference(v, w, disp = 2)

}
