% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_traits.R
\name{unscale_singleopt}
\alias{unscale_singleopt}
\title{Back transform output to original scale}
\usage{
unscale_singleopt(trait_list, singleopt_output)
}
\arguments{
\item{trait_list}{Original list of trait data as used in `singleopt_context`.}

\item{singleopt_output}{Archive output from `singleopt_context`.}
}
\value{
List with archive values and weights matching formatting of `singleopt_output`.
}
\description{
This function uses the original transformed trait data provided to `singleopt_context` and the corresponding output
from `singleopt_context` to back transform archive values to the original trait scale.
This function assumes trait data was transformed using `scale_traits`.
}
\details{
Note that this function takes the absolute value of transformed trait variables.
}
