% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_measure.R
\name{wasserstein_measure}
\alias{wasserstein_measure}
\title{Calculate the Wasserstein Distance for use in simulated annealing}
\usage{
wasserstein_measure(v, w, ...)
}
\arguments{
\item{v}{single-column matrix of numeric values. Will be coerced to a vector.}

\item{w}{Numeric vector of individual weights with same length as `v`.}

\item{...}{Optional additional arguments. See \code{?wasserstein} for specific
options.}
}
\value{
A single numeric value representing the p-Wasserstein distance multiplied
by -1.
}
\description{
Uses the \code{wasserstein()} to calculate the Wasserstein distance (\eqn{W_p}) between
the original trait data distribution and the proposed subset of individuals.
Assumes that v provides the total population distribution of trait variation.
}
\details{
As `MultiOpt` aims to maximize measures, but smaller \eqn{W_p} are
better, we multiply the \eqn{W_p} by -1 to allow for minimizing
the differences in distributions.


This function examines how well distributions match. If you are more interested
in checking the full range of trait data is captured in the subset,
\code{trait_coverage()} would be a better fit.
}
\examples{
v <- as.matrix(c(1, 2, 3))
w <- c(0, 1, 1)
wasserstein_measure(v, w)
}
